Abstract pipe
