+++ src/main/scala/com/kotori316/fluidtank/network/ClientProxy.java
@@ -65,7 +65,7 @@ public class ClientProxy extends SideProxy {
         public static void registerTESR(FMLClientSetupEvent event) {
             ClientRegistry.bindTileEntityRenderer(ModObjects.TANK_TYPE(), RenderTank::new);
             ClientRegistry.bindTileEntityRenderer(ModObjects.TANK_CREATIVE_TYPE(), RenderTank::new);
-            ClientRegistry.bindTileEntityRenderer(ModObjects.PIPE_TYPE(), d -> {
+            ClientRegistry.bindTileEntityRenderer(ModObjects.FLUID_PIPE_TYPE(), d -> {
                 RenderPipe renderPipe = new RenderPipe(d);
                 renderPipe.useColor_$eq(!Config.content().enablePipeRainbowRenderer().get());
                 return renderPipe;
@@ -77,7 +77,7 @@ public class ClientProxy extends SideProxy {
                 .forEach(tank -> RenderTypeLookup.setRenderLayer(tank, rendertype));
             CollectionConverters.asJava(ModObjects.blockTanksInvisible())
                 .forEach(tank -> RenderTypeLookup.setRenderLayer(tank, rendertype));
-            RenderTypeLookup.setRenderLayer(ModObjects.blockPipe(), rendertype);
+            RenderTypeLookup.setRenderLayer(ModObjects.blockFluidPipe(), rendertype);
         }
 
         @SubscribeEvent
